+++ CHANGELOG.md
@@ -60,6 +60,8 @@ The git repositories are hosted at the following sites:
 
 ### Fixed
 
+* Fixed basic network port object to allow clearing of previously
+  set remote BBMD address in Network_Port_FD_BBMD_Address_Write. (#1482)
 * Fixed leaky memory management of Keylist based List properties
   in the basic Audit Log, Load Control, Life Safety Zone,
   Multistate, Network Port, and Structured View objects (#1479)